api/requirements.txt
kafka-python==2.0.2
beanie==1.10.8
motor==2.5.1
pydantic==1.9.0
fastapi==0.75.2
uvicorn[standard]==0.17.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
sse-starlette==0.1.0
